target: Add continuous splash display implementation
Most devices leave the panel powered on to keep the splash screen
visible while the kernel is loading. This allows having a very
simple display implementation in the secondary bootloader by using
the memory region that was allocated for the continuous splash.

The concept is borrowed from efidroid:
https://github.com/efidroid/bootloader_lk/tree/common/shared/lib/2ndstage_display
but the code has changed quite a bit to simplify it.

# Conflicts:
#	target/rules.mk